refactor: reviewed and updated key ID naming conventions across models and migrations to align with Laravel best practices

+++ database/migrations/2025_06_20_184413_add_franchisee_id_and_created_by_and_updated_by_in_additionalcharges_table.php
@@ -4,21 +4,22 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('additionalcharges', function (Blueprint $table) {
-            $table->unsignedBigInteger('franchisee_id')->nullable();
-            $table->foreign('franchisee_id')->references('franchisee_id')->on('franchisees')->onDelete('set null');
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
-            $table->foreign('created_by')->references('user_id')->on('users')->onDelete('set null');
-            $table->foreign('updated_by')->references('user_id')->on('users')->onDelete('set null');
-        });
+        if (Schema::hasTable('additionalcharges') && !Schema::hasColumn('additionalcharges', 'franchise_id')) {
+            Schema::table('additionalcharges', function (Blueprint $table) {
+                $table->unsignedBigInteger('franchise_id')->nullable();
+                $table->foreign('franchise_id')->references('user_id')->on('users')->onDelete('set null');
+                $table->unsignedBigInteger('created_by')->nullable();
+                $table->unsignedBigInteger('updated_by')->nullable();
+                $table->foreign('created_by')->references('user_id')->on('users')->onDelete('set null');
+                $table->foreign('updated_by')->references('user_id')->on('users')->onDelete('set null');
+            });
+        }
     }
 
     /**
@@ -26,11 +27,13 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('additionalcharges', function (Blueprint $table) {
-            $table->dropForeign(['franchisee_id']);
-            $table->dropForeign(['created_by']);
-            $table->dropForeign(['updated_by']);
-            $table->dropColumn(['franchisee_id', 'created_by', 'updated_by']);
-        });
+        if (Schema::hasTable('additionalcharges') && Schema::hasColumn('additionalcharges', 'franchise_id')) {
+            Schema::table('additionalcharges', function (Blueprint $table) {
+                $table->dropForeign(['franchise_id']);
+                $table->dropForeign(['created_by']);
+                $table->dropForeign(['updated_by']);
+                $table->dropColumn(['franchise_id', 'created_by', 'updated_by']);
+            });
+        }
     }
 };
